# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ILLICO DELIVERY API",
    "version": "1.0.0",
    "description": "Plateforme de livraison express & points de collecte — Cotonou & Abomey-Calavi\n\n## 🎯 Fonctionnalités\n- Authentification multi-rôles (Admin, Client, Livreur, Point ILLICO)\n- Gestion des livraisons en temps réel\n- Points de collecte avec stockage temporaire\n- Calcul tarifaire dynamique\n- Paiements & commissions\n- Notifications push\n\n## 🔐 Authentification\nToutes les routes protégées nécessitent un token JWT dans le header :\n```\nAuthorization: Bearer <votre_token_jwt>\n```\n\n## 🌍 Base URL\n- Development: `http://localhost:3000/api`\n- Production: `https://illico-delivery.vercel.app/api`"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development (Local)"
    },
    {
      "url": "https://illico-delivery.vercel.app/api",
      "description": "Production (Vercel)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenu après connexion. Format: `Bearer <token>`"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "description": "Utilisateur (4 rôles possibles: Admin, Client, Livreur, PointIllico)",
        "properties": {
          "_id": { "type": "string", "example": "65f2a1b2c3d4e5f6a7b8c9d0" },
          "nom": { "type": "string", "example": "Koffi Mensah" },
          "telephone": { "type": "string", "example": "+22997123456" },
          "email": { "type": "string", "example": "admin@illicodelivery.bj", "nullable": true },
          "role": { "type": "string", "enum": ["Admin", "Client", "Livreur", "PointIllico"], "example": "Client" },
          "adresse": { "type": "string", "example": "Cotonou, Haie Vive, Rue 123", "nullable": true },
          "typeClient": { "type": "string", "enum": ["standard", "professionnel"], "example": "professionnel", "nullable": true },
          "soldeIllico": { "type": "number", "example": 15000, "nullable": true },
          "forfaitActif": { "type": "string", "nullable": true, "example": "65f2a1b2c3d4e5f6a7b8c9d1" },
          "vehicule": { "type": "object", "nullable": true, "example": { "_id": "65f2a1b2c3d4e5f6a7b8c9d2", "type": "moto", "tarifBase": 500 } },
          "statut": { "type": "string", "enum": ["en_ligne", "hors_ligne", "en_mission"], "example": "en_ligne", "nullable": true },
          "scoreNote": { "type": "number", "minimum": 0, "maximum": 5, "example": 4.8, "nullable": true },
          "cashCollecte": { "type": "number", "example": 25000, "nullable": true },
          "cashReverse": { "type": "number", "example": 10000, "nullable": true },
          "valide": { "type": "boolean", "example": true, "nullable": true },
          "actif": { "type": "boolean", "example": true, "nullable": true },
          "commissionTotal": { "type": "number", "example": 5000, "nullable": true },
          "location": {
            "type": "object",
            "properties": {
              "type": { "type": "string", "example": "Point" },
              "coordinates": { "type": "array", "items": { "type": "number" }, "example": [2.4394, 6.3654] }
            }
          }
        }
      },
      "Livraison": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "example": "65f2a1b2c3d4e5f6a7b8c9e0" },
          "client": { "$ref": "#/components/schemas/User" },
          "livreur": { "$ref": "#/components/schemas/User", "nullable": true },
          "pointDepart": {
            "type": "object",
            "properties": {
              "adresse": { "type": "string", "example": "Cotonou, Haie Vive" },
              "coordinates": { "type": "array", "items": { "type": "number" }, "example": [2.4394, 6.3654] }
            }
          },
          "pointArrivee": {
            "type": "object",
            "properties": {
              "adresse": { "type": "string", "example": "Abomey-Calavi, Godomey" },
              "coordinates": { "type": "array", "items": { "type": "number" }, "example": [2.3558, 6.4478] }
            }
          },
          "vehicule": { "type": "object", "properties": { "_id": { "type": "string" }, "type": { "type": "string" }, "tarifBase": { "type": "number" } } },
          "mode": { "type": "string", "enum": ["express", "point_illico"], "example": "express" },
          "pointIllico": { "$ref": "#/components/schemas/User", "nullable": true },
          "statut": { "type": "string", "enum": ["en_attente", "affecté", "arrivé_pickup", "colis_récupéré", "déposé_en_point", "livré", "échoué", "annulé"], "example": "en_attente" },
          "prixEstime": { "type": "number", "example": 2500 },
          "prixFinal": { "type": "number", "nullable": true, "example": 2500 },
          "modePaiement": { "type": "string", "enum": ["cash", "mobile_money", "credit_illico"], "example": "cash" },
          "otpLivraison": { "type": "string", "example": "123456", "writeOnly": true },
          "otpRetrait": { "type": "string", "example": "789012", "writeOnly": true },
          "otpValidé": { "type": "boolean", "example": false },
          "preuveLivraisonUrl": { "type": "string", "example": "https://res.cloudinary.com/.../proof.jpg", "nullable": true },
          "noteLivreur": { "type": "number", "minimum": 1, "maximum": 5, "nullable": true },
          "urgent": { "type": "boolean", "example": false },
          "nuit": { "type": "boolean", "example": false },
          "poids": { "type": "number", "example": 2.5 },
          "dateCreation": { "type": "string", "format": "date-time", "example": "2026-04-15T10:30:00Z" },
          "dateLivraison": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "Colis": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "example": "65f2a1b2c3d4e5f6a7b8c9f0" },
          "reference": { "type": "string", "example": "COL-K9X2M-4P7Q" },
          "livraison": { "$ref": "#/components/schemas/Livraison" },
          "pointIllico": { "$ref": "#/components/schemas/User" },
          "statut": { "type": "string", "enum": ["en_attente", "receptionné", "retiré", "retourné"], "example": "receptionné" },
          "dateDepot": { "type": "string", "format": "date-time", "example": "2026-04-15T14:00:00Z" },
          "dateLimiteGratuite": { "type": "string", "format": "date-time", "example": "2026-04-17T14:00:00Z" },
          "dateRetrait": { "type": "string", "format": "date-time", "nullable": true },
          "fraisStockage": { "type": "number", "example": 0 },
          "fraisStockageActuel": { "type": "number", "example": 400, "description": "Calculé à la volée si retard" },
          "otpRetrait": { "type": "string", "example": "456789", "writeOnly": true },
          "signatureUrl": { "type": "string", "nullable": true }
        }
      },
...
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/colis/{id}": {
      "get": {
        "tags": ["📦 Colis"],
        "summary": "Détail d'un colis",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Colis récupéré",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/Colis" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}